requirements.txt
pytrends
google-generativeai
requests
python-dotenv
